Merges `next` into `merge-train/spartan` to unblock the daily auto-sync.

The `merge-train-next-to-branches` workflow runs `merge-next.sh
merge-train/spartan || true`, so when the merge hits a conflict it
aborts but the workflow still reports success — which is why the train
has been stuck a day behind `next`.

The only conflict is in `docs/docs/networks.md`: the train carried a
stale testnet **Slash Factory** address (from #24648, the v5
rollup-upgrade docs update), while `next` has since set it to `N/A`.
Resolved in favour of `next` (the newer state); this file now matches
`origin/next` exactly. No other files conflicted.

Per the merge-train conflict procedure, this should be force-merged with
`ci-skip` using the **merge commit** method (not squash) to preserve the
resolution.

+++ barretenberg/cpp/src/barretenberg/polynomials/eq_polynomial.test.cpp
@@ -207,7 +207,7 @@ TEST_F(EqPolyTest, ProverTableMatchesVerifierOnBooleanPoints)
     for (uint64_t ell = 0; ell < (1ULL << d); ++ell) {
         const auto u = bool_vec_from_mask(d, ell);
         const FF got_ver = v.evaluate(u);
-        const FF got_prov = peq.at(ell);
+        const FF got_prov = peq.at(static_cast<size_t>(ell));
         EXPECT_EQ(got_prov, got_ver) << "ell=" << ell;
     }
 }
